server/requirements.txt
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.38.0
contourpy==1.2.0
cvzone==1.6.1
cycler==0.12.1
dnspython==2.5.0
filelock==3.13.1
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
fonttools==4.46.0
fsspec==2023.12.2
geographiclib==2.0
geopy==2.4.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
nest-asyncio==1.5.8
networkx==3.2.1
numpy==1.26.2
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
psutil==5.9.6
py-cpuinfo==9.0.0
PyJWT==2.8.0
pymongo==4.6.1
pyngrok==7.0.3
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
scipy==1.11.4
seaborn==0.13.0
six==1.16.0
SQLAlchemy==2.0.25
sympy==1.12
thop==0.1.1.post2209072238
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.3
ultralytics==8.0.227
urllib3==2.1.0
Werkzeug==3.0.1
flask
flask-cors
flask-jwt-extended
pymongo
python-dotenv
werkzeug

